+++ Changelog.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## [Unreleased]
-
+## [3.16.1] - 2026-09-02
+### Added
+- Added `DELETE_WIDGET_CONFIG` access permission for deleting widget configurations.
+- Added optional `linkedWidgetConfigIds` parameter to store linked widget configuration IDs.
+- Updated customer onboarding to pass linkedWidgetConfigIds to the KYC verifier page and name/description when configuring the ID widget.
 
 ## [3.16.0] - 2026-08-27
 ### Added